requirements.txt
# CLI
cleo==0.8.1
colorlog==4.8.0

# YAML
PyYAML==5.4.1

# Clien Web
aiohttp==3.7.4

# Asyncio
uvloop==0.15.2

# Database
tortoise-orm[asyncpg]==0.17.2

# Scraper
lxml==4.6.3
